backend/requirements.txt
# Django and DRF
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.11.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-allauth>=0.54.0

# Redis & Caching
redis>=4.5.0
django-redis>=5.3.0

# Development & Utilities
python-decouple>=3.8
Pillow>=9.5.0
django-filter>=23.0
django-extensions>=3.2.0

# Production
gunicorn>=20.1.0
whitenoise>=6.5.0
